Rename consts.c to consts.h, use enums in const.h in Main.c
